# Use System.Text.Json instead of Newton.Json