--- a/AppInspector.Benchmarks/AppInspector.Benchmarks.csproj
+++ b/AppInspector.Benchmarks/AppInspector.Benchmarks.csproj
@@ -5,6 +5,8 @@
         <TargetFrameworks>net6.0;net7.0</TargetFrameworks>
         <Nullable>enable</Nullable>
         <LangVersion>10.0</LangVersion>
+        <RootNamespace>ApplicationInspector.Benchmarks</RootNamespace>
+        <Company>Microsoft Corporation</Company>
     </PropertyGroup>
 
     <ItemGroup>