--- a/AppInspector.RulesEngine/AppInspector.RulesEngine.csproj
+++ b/AppInspector.RulesEngine/AppInspector.RulesEngine.csproj
@@ -35,7 +35,6 @@
         <PackageReference Include="Microsoft.CST.OAT" Version="1.2.42" />
         <PackageReference Include="Microsoft.CST.RecursiveExtractor" Version="1.2.8" />
         <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.0" />
-        <PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
         <PackageReference Include="YamlDotNet" Version="12.3.1" />
     </ItemGroup>