--- a/AppInspector.RulesEngine/Comment.cs
+++ b/AppInspector.RulesEngine/Comment.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright(C) Microsoft.All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-using Newtonsoft.Json;
+using System.Text.Json.Serialization;
 
 namespace Microsoft.ApplicationInspector.RulesEngine;
 
@@ -10,15 +10,15 @@ namespace Microsoft.ApplicationInspector.RulesEngine;
 /// </summary>
 internal class Comment
 {
-    [JsonProperty(PropertyName = "language")]
+    [JsonPropertyName("language")]
     public string[]? Languages { get; set; }
 
-    [JsonProperty(PropertyName = "inline")]
+    [JsonPropertyName("inline")]
     public string? Inline { get; set; }
 
-    [JsonProperty(PropertyName = "prefix")]
+    [JsonPropertyName("prefix")]
     public string? Prefix { get; set; }
 
-    [JsonProperty(PropertyName = "suffix")]
+    [JsonPropertyName("suffix")]
     public string? Suffix { get; set; }
 }