--- a/AppInspector.RulesEngine/LanguageInfo.cs
+++ b/AppInspector.RulesEngine/LanguageInfo.cs
@@ -1,8 +1,7 @@
 ﻿// Copyright(C) Microsoft.All rights reserved.
 // Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
+using System.Text.Json.Serialization;
 
 namespace Microsoft.ApplicationInspector.RulesEngine;
 
@@ -17,15 +16,16 @@ public class LanguageInfo
         Build
     }
 
-    [JsonProperty(PropertyName = "name")] public string Name { get; set; } = "";
+    [JsonPropertyName("name")]
+    public string Name { get; set; } = "";
 
-    [JsonProperty(PropertyName = "extensions")]
+    [JsonPropertyName("extensions")]
     public string[]? Extensions { get; set; }
 
-    [JsonProperty(PropertyName = "file-names")]
+    [JsonPropertyName("file-names")]
     public string[]? FileNames { get; set; }
 
-    [JsonProperty(PropertyName = "type")]
-    [JsonConverter(typeof(StringEnumConverter))]
+    [JsonPropertyName("type")]
+    [JsonConverter(typeof(JsonStringEnumConverter))]
     public LangFileType Type { get; set; } = LangFileType.Code;
 }