… AppInspector.Benchmarks/AnalyzeBenchmark.cs
… AppInspector.Benchmarks/DistinctBenchmarks.cs
… AppInspector.Benchmarks/Program.cs
… AppInspector.RulesEngine/MatchRecord.cs
… AppInspector.RulesEngine/PatternScope.cs
… AppInspector.RulesEngine/PatternType.cs
… AppInspector.RulesEngine/Rule.cs
… AppInspector.RulesEngine/SearchCondition.cs
… AppInspector.RulesEngine/SearchPattern.cs
… AppInspector.RulesEngine/Severity.cs
… AppInspector.RulesEngine/TypedRuleSet.cs
… AppInspector.Tests/AppInspector.Tests.csproj
… AppInspector.Tests/Languages/LanguagesTests.cs
… AppInspector.Tests/Writers/TestVerifyRulesCmd.cs
… AppInspector.sln
… AppInspector/AppInspector.Commands.csproj
… AppInspector/Commands/AnalyzeCommand.cs
… AppInspector/Commands/ExportTagsCommand.cs
… AppInspector/Commands/PackRulesCommand.cs
… AppInspector/Commands/TagDiffCommand.cs
… AppInspector/Commands/VerifyRulesCommand.cs
… AppInspector/FileRecord.cs
… AppInspector/MetaData.cs
… AppInspector/MetaDataHelper.cs
… AppInspector/MetricTagCounter.cs
… AppInspector/Result.cs
… Benchmarks/Benchmarks.csproj
… Benchmarks/Program.cs
… Benchmarks/WriterBench.cs
… version.json