Fixes #515