@gfs 2023-02-08
Thanks for your contribution.  This is an excellent improvement I think.

I added a simple benchmark to test the speed of just the export time. It generates a trivial report with N issues for setup and then exports the same report to JSON repeatedly. The benchmark shows about a 2x speedup, and about 10x less memory allocations.

Current main branch
|                Method |     N |      Mean |     Error |    StdDev | Ratio |     Gen0 |  Allocated | Alloc Ratio |
|---------------------- |------ |----------:|----------:|----------:|------:|---------:|-----------:|------------:|
| ExportRecordsToJson |  1000 |  3.680 ms | 0.0736 ms | 0.1833 ms |  1.00 |  23.4375 |  387.88 KB |        1.00 |
|                       |       |           |           |           |       |          |            |             |
| ExportRecordsToJson | 10000 | 30.532 ms | 0.6079 ms | 1.5138 ms |  1.00 | 218.7500 | 3762.89 KB |        1.00 |

This PR
|                Method |     N |      Mean |     Error |    StdDev | Ratio |    Gen0 | Allocated | Alloc Ratio |
|---------------------- |------ |----------:|----------:|----------:|------:|--------:|----------:|------------:|
| ExportRecordsToJson |  1000 |  2.052 ms | 0.0494 ms | 0.1442 ms |  1.00 |       - |  52.17 KB |        1.00 |
|                       |       |           |           |           |       |         |           |             |
| ExportRecordsToJson | 10000 | 13.035 ms | 0.2579 ms | 0.4310 ms |  1.00 | 15.6250 |  480.6 KB |        1.00 |