@gfs 2023-02-08
@dbalikhin I did a quick check of the json, sarif and html outputs compared to the latest release from main and everything looks good, as well as double checking that custom rules can still be specified and loaded properly.

Thanks again for the contribution, great work!